requirements.txt
python-dotenv~=0.20
psycopg2~=2.9

Django~=4.1
django-cors-headers~=3.13

djangorestframework~=3.13

#GDAL>=3.2.0,<4.0.0

flake8~=5.0

django-cors-headers~=3.13
djangorestframework-gis~=1.0

django-filter~=23.1
django-modeltranslation~=0.18.9